pull out routing helpers into separate file
